refactor: graph에 관한 컴포넌트와 class를  분리
