+++  README.md
@@ -31,3 +31,10 @@
 
 - [x] 최소 시간 경로 기능을 추가
 - [x] 최단 경로 조회 기능과의 중복 제거를 수행(테스트 코드도 마찬가지로 중복제거)
+
+## TODO
+- [ ] graph 클래스 분리
+- [ ] line 도메인 일급 컬렉션 및 list로 변경 경 
+- [ ] exception 추상화
+- [ ] exception handler 추가
+- [ ] javascript 리팩토링
