+++ src/main/java/wooteco/subway/admin/controller/restcontroller/LineController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.controller;
+package wooteco.subway.admin.controller.restcontroller;
 
 import java.net.URI;
 import java.util.List;
@@ -15,10 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.dto.request.LineRequest;
-import wooteco.subway.admin.dto.request.LineStationCreateRequest;
-import wooteco.subway.admin.dto.response.LineDetailResponse;
 import wooteco.subway.admin.dto.response.LineResponse;
-import wooteco.subway.admin.dto.response.WholeSubwayResponse;
 import wooteco.subway.admin.service.LineService;
 
 @RestController
@@ -44,11 +41,6 @@ public class LineController {
         return ResponseEntity.ok().body(lineService.showLines());
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<LineDetailResponse> retrieveLine(@PathVariable Long id) {
-        return ResponseEntity.ok().body(lineService.findLineWithStationsById(id));
-    }
-
     @PutMapping("/{id}")
     public ResponseEntity updateLine(@PathVariable Long id, @RequestBody LineRequest view) {
         lineService.updateLine(id, view);
@@ -58,26 +50,6 @@ public class LineController {
     @DeleteMapping("/{id}")
     public ResponseEntity deleteLine(@PathVariable Long id) {
         lineService.deleteLineById(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/detail")
-    public ResponseEntity<WholeSubwayResponse> retrieveWholeSubway() {
-        WholeSubwayResponse wholeSubwayResponse = lineService.wholeLines();
-        return ResponseEntity.ok().body(wholeSubwayResponse);
-    }
-
-    @PostMapping("/{lineId}/stations")
-    public ResponseEntity addLineStation(@PathVariable Long lineId,
-        @RequestBody LineStationCreateRequest view) {
-        lineService.addLineStation(lineId, view);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{lineId}/stations/{stationId}")
-    public ResponseEntity removeLineStation(@PathVariable Long lineId,
-        @PathVariable Long stationId) {
-        lineService.removeLineStation(lineId, stationId);
         return ResponseEntity.noContent().build();
     }
 }
