+++ src/main/java/wooteco/subway/admin/dto/response/PathResponse.java
@@ -16,6 +16,10 @@ public class PathResponse {
         this.distance = distance;
     }
 
+    public static PathResponse of(List<StationResponse> stations, int duration, int distance) {
+        return new PathResponse(stations, duration, distance);
+    }
+
     public List<StationResponse> getStations() {
         return stations;
     }
