+++ src/main/java/wooteco/subway/admin/service/LineService.java
@@ -1,98 +1,42 @@
 package wooteco.subway.admin.service;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.LineStation;
-import wooteco.subway.admin.domain.Station;
 import wooteco.subway.admin.dto.request.LineRequest;
-import wooteco.subway.admin.dto.request.LineStationCreateRequest;
-import wooteco.subway.admin.dto.response.LineDetailResponse;
 import wooteco.subway.admin.dto.response.LineResponse;
-import wooteco.subway.admin.dto.response.WholeSubwayResponse;
 import wooteco.subway.admin.repository.LineRepository;
-import wooteco.subway.admin.repository.StationRepository;
 
 @Service
 public class LineService {
     private LineRepository lineRepository;
-    private StationRepository stationRepository;
 
-    public LineService(LineRepository lineRepository, StationRepository stationRepository) {
+    public LineService(LineRepository lineRepository) {
         this.lineRepository = lineRepository;
-        this.stationRepository = stationRepository;
     }
 
-    public Line save(Line line) {
-        return lineRepository.save(line);
-    }
-
+    @Transactional(readOnly = true)
     public List<LineResponse> showLines() {
         return LineResponse.listOf(lineRepository.findAll());
     }
 
+    @Transactional
+    public Line save(Line line) {
+        return lineRepository.save(line);
+    }
+
+    @Transactional
     public void updateLine(Long id, LineRequest request) {
         Line persistLine = lineRepository.findById(id).orElseThrow(RuntimeException::new);
         persistLine.update(request.toLine());
         lineRepository.save(persistLine);
     }
 
+    @Transactional
     public void deleteLineById(Long id) {
         lineRepository.deleteById(id);
-    }
-
-    public void addLineStation(Long id, LineStationCreateRequest request) {
-        validateStations(request.getPreStationId(), request.getStationId());
-        Line line = lineRepository.findById(id).orElseThrow(RuntimeException::new);
-        LineStation lineStation = LineStation.of(request.getPreStationId(), request.getStationId(),
-            request.getDistance(), request.getDuration());
-        line.addLineStation(lineStation);
-
-        lineRepository.save(line);
-    }
-
-    private void validateStations(Long preStationId, Long stationId) {
-        if (Objects.nonNull(preStationId) && !stationRepository.existsById(preStationId)) {
-            throw new IllegalArgumentException("존재하지 않는 이전역입니다.");
-        }
-        if (Objects.isNull(stationId) || !stationRepository.existsById(stationId)) {
-            throw new IllegalArgumentException("존재하지 않는 현재역입니다.");
-        }
-    }
-
-    public void removeLineStation(Long lineId, Long stationId) {
-        Line line = lineRepository.findById(lineId).orElseThrow(RuntimeException::new);
-        line.removeLineStationById(stationId);
-        lineRepository.save(line);
-    }
-
-    public LineDetailResponse findLineWithStationsById(Long id) {
-        Line line = lineRepository.findById(id).orElseThrow(RuntimeException::new);
-        List<Station> stations = stationRepository.findAllById(line.getLineStationsId());
-        return LineDetailResponse.of(line, stations);
-    }
-
-    public WholeSubwayResponse wholeLines() {
-        List<Line> lines = lineRepository.findAll();
-        Map<Long, Station> stations = stationRepository.findAll()
-            .stream()
-            .collect(Collectors.toMap(Station::getId, station -> station));
-        List<LineDetailResponse> responses = lines.stream()
-            .map(line -> getLineDetailResponse(stations, line))
-            .collect(Collectors.toList());
-        return WholeSubwayResponse.of(responses);
-    }
-
-    private LineDetailResponse getLineDetailResponse(Map<Long, Station> stations, Line line) {
-        List<Long> stationIds = line.getLineStationsId();
-        List<Station> stationsList = stationIds.stream()
-            .map(stations::get)
-            .collect(Collectors.toList());
-        return LineDetailResponse.of(line, stationsList);
     }
 }
