+++ src/main/resources/templates/admin/admin-station.html
@@ -1,16 +1,16 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    <meta charset="UTF-8"/>
-    <title>subway admin</title>
-    <meta
-      content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, shrink-to-fit=no"
-      name="viewport"/>
-    <link href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css" rel="stylesheet"/>
-    <link href="https://cdn.materialdesignicons.com/5.0.45/css/materialdesignicons.min.css"
-          rel="stylesheet"/>
-    <link href="/admin/css/admin-app.css" rel="stylesheet"/>
-    <link href="/admin/css/admin-station.css" rel="stylesheet"/>
+  <meta charset="UTF-8"/>
+  <title>subway admin</title>
+  <meta
+    content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, shrink-to-fit=no"
+    name="viewport"/>
+  <link href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css" rel="stylesheet"/>
+  <link href="https://cdn.materialdesignicons.com/5.0.45/css/materialdesignicons.min.css"
+        rel="stylesheet"/>
+  <link href="/admin/css/admin-app.css" rel="stylesheet"/>
+  <link href="/admin/css/admin-station.css" rel="stylesheet"/>
 </head>
 <body class="bg-gray-200 h-screen">
 <div class="flex justify-center md:py-10 lg:py-10 app-container">
@@ -34,17 +34,17 @@
       </div>
     </form>
     <div class="mt-5 flex station-list-container relative overflow-y-auto">
-        <div class="w-full" id="station-list">
-            {{#each stations}}
-            <div class="list-item border border-gray-200 py-2 px-4 text-gray-800" data-id="{{id}}">
-                {{name}}
-                <button
-                  class="hover:bg-gray-300 hover:text-gray-500 text-gray-300 px-1 rounded-full float-right">
-                    <span class="mdi mdi-delete"></span>
-                </button>
-            </div>
-            {{/each}}
+      <div class="w-full" id="station-list">
+        {{#each stations}}
+        <div class="list-item border border-gray-200 py-2 px-4 text-gray-800" data-id="{{id}}">
+          {{name}}
+          <button
+            class="hover:bg-gray-300 hover:text-gray-500 text-gray-300 px-1 rounded-full float-right">
+            <span class="mdi mdi-delete"></span>
+          </button>
         </div>
+        {{/each}}
+      </div>
     </div>
   </div>
 </div>
